bot/requirements.txt
python-dotenv==1.1.1
groq==0.30.0
flask==3.1.1
flask-cors==6.0.1
aiohttp==3.12.15
python-telegram-bot==22.3
gunicorn==23.0.0
